requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
drf-yasg==1.21.10
PyMySQL==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
